added a couple of tests, fixed IndPSD?

+++ src/functions/normL21.jl
@@ -10,8 +10,8 @@ In words, it is the (weighted) sum of the Euclidean norm of the columns (rows) o
 
 immutable NormL21 <: NormFunction
   lambda::Real
-  dim::Int
-  NormL21(lambda::Real=1.0, dim=1) =
+  dim::Integer
+  NormL21(lambda::Real=1.0, dim::Integer=1) =
     lambda < 0 ? error("parameter λ must be nonnegative") : new(lambda, dim)
 end
